Moved file structure around. Added method documentation and JSON logger.
