+++ src/Logger/DeadLogger.ts
@@ -1,4 +1,4 @@
-import { LoggerStructure, LogLevel, LogFormatterStructure, Stream } from "./compiler/types";
+import { LoggerStructure, LogLevel, LogFormatterStructure, Stream } from "../compiler/types";
 
 export default class DeadLogger implements LoggerStructure {
     trace(message: string | object | Error, ...args: any[]): void { return; }
@@ -9,9 +9,11 @@ export default class DeadLogger implements LoggerStructure {
     fatal(message: string | object | Error, ...args: any[]): void { return; }
     log(level: LogLevel, message: string | object | Error, ...args: any[]): void { return; }
     child(properties: object): LoggerStructure { return this; }
-    addLogProperty(key: string, value: any): void { return; }
+    addLogProperty(key: string, value: string | object): void { return; }
     removeLogProperty(key: string): void { return; }
-    attachPropertyScope(key: string, property: string | object, cb: () => any) { return; }
+    scope(key: string, property: string | object, cb: () => any) { return; }
+    async asyncScope(key: string, property: string | object, cb: () => Promise<Boolean>) { return true; }
     setFormatter(formatter: LogFormatterStructure): LoggerStructure { return this; }
     addStream(stream: Stream): LoggerStructure { return this; }
+    mask(input: string | number | undefined): string { return ""; }
 }
