+++ src/index.ts
@@ -1,5 +1,7 @@
-import Logger from './Logger';
+import Logger from './Logger/Logger';
 import { LoggerStructure, LogFormatterStructure, LogLevel, LogMessage, Stream, LogProperties, LogEvent, OutputStream } from './compiler/types';
+import LinearFormatter from './formatters/LinearFormatter';
+import JSONFormatter from './formatters/JSONFormatter';
 
 export {
     LoggerStructure,
@@ -10,6 +12,8 @@ export {
     LogProperties,
     LogEvent,
     OutputStream,
+    LinearFormatter,
+    JSONFormatter
 };
 
 export default Logger;
